fusemap: Support different word sizes for NVMEM device and fusemap description
In case of the i.MX93 the NVMEM driver in the mainline kernel
uses a word size of 1 byte but the reference manual uses the
hardware scheme of 4 byte words as used with i.MX6 and i.MX8.

In order to describe the registers in the YAML in accordance
with the hardware and still get the correct addresses when
accessing the NVMEM device, let's use separat parameters
for the word size in the fusemap description and the NVMEM
driver.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>